Simplify JSON schema and only generate appsettings-schema.Umbraco.Cms.json
