--- src/JsonSchema/NamespacePrefixedSchemaNameGenerator.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) Umbraco.
-// See LICENSE for more details.
-
-using System;
-using NJsonSchema.Generation;
-
-namespace JsonSchema
-{
-    internal class NamespacePrefixedSchemaNameGenerator : DefaultSchemaNameGenerator
-    {
-        public override string Generate(Type type) => type.Namespace?.Replace(".", string.Empty) + base.Generate(type);
-    }
-}
