--- src/JsonSchema/appsettings-schema.json
@@ -1,61 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "properties": {
-    "Umbraco": {
-      "description": "The container of all Umbraco content",
-      "oneOf": [
-        {
-          "type": "null"
-        },
-        {
-          "properties": {
-            "CMS": {
-              "description": "Configuration of Umbraco CMS",
-              "oneOf": [
-                {
-                  "type": "null"
-                },
-                {
-                  "$ref": "appsettings-schema.Umbraco.Cms.json#/definitions/JsonSchemaCmsDefinition"
-                }
-              ]
-            },
-            "Forms": {
-              "description": "Configuration of Umbraco Forms",
-              "oneOf": [
-                {
-                  "type": "null"
-                },
-                {
-                  "$ref": "appsettings-schema.Umbraco.Forms.json#/definitions/JsonSchemaFormsDefinition"
-                }
-              ]
-            },
-            "Deploy": {
-              "description": "Configuration of Umbraco Deploy",
-              "oneOf": [
-                {
-                  "type": "null"
-                },
-                {
-                  "$ref": "appsettings-schema.Umbraco.Deploy.json#/definitions/JsonSchemaDeployDefinition"
-                }
-              ]
-            },
-            "Workflow": {
-              "description": "Configuration of Umbraco Workflow",
-              "oneOf": [
-                {
-                  "type": "null"
-                },
-                {
-                  "$ref": "appsettings-schema.Umbraco.Workflow.json#/definitions/JsonSchemaWorkflowDefinition"
-                }
-              ]
-            }
-          }
-        }
-      ]
-    }
-  }
-}
