+++ src/Umbraco.Cms.Targets/appsettings-schema.json
@@ -1,0 +1,11 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "allOf": [
+    {
+      "$ref": "https://json.schemastore.org/appsettings.json"
+    },
+    {
+      "$ref": "appsettings-schema.Umbraco.Cms.json#"
+    }
+  ]
+}
