+++ src/Umbraco.Web.UI/appsettings.Development.template.json
@@ -1,5 +1,5 @@
 {
-  "$schema" : "./appsettings-schema.json",
+  "$schema": "appsettings-schema.json",
   "Serilog": {
     "MinimumLevel": {
       "Default": "Information",
