+++ src/Umbraco.Web.UI/appsettings.template.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "./appsettings-schema.json",
+  "$schema": "appsettings-schema.json",
   "ConnectionStrings": {
     "umbracoDbDSN": ""
   },
