+++ templates/UmbracoProject/appsettings.Development.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "./appsettings-schema.json",
+  "$schema": "appsettings-schema.json",
   "Serilog": {
     "MinimumLevel": {
       "Default": "Information"
