+++ tools/Umbraco.JsonSchema/Options.cs
@@ -1,0 +1,7 @@
+using CommandLine;
+
+internal class Options
+{
+    [Option("outputFile", Default = "appsettings-schema.Umbraco.Cms.json", HelpText = "Output file to save the generated JSON schema for Umbraco CMS.")]
+    public string OutputFile { get; set; } = null!;
+}
